rename application to felfele in android and ios
