+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.postmodern",
+    package = "org.felfele.mobile",
 )
 
 android_resource(
     name = "res",
-    package = "com.postmodern",
+    package = "org.felfele.mobile",
     res = "src/main/res",
 )
