+++ android/app/build.gradle
@@ -96,7 +96,7 @@ android {
     buildToolsVersion "27.0.3"
 
     defaultConfig {
-        applicationId "com.postmodern"
+        applicationId "org.felfele.mobile"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
