+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'postmodern'
+rootProject.name = 'Felfele'
 include ':react-native-background-fetch'
 project(':react-native-background-fetch').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-background-fetch/android')
 include ':react-native-push-notification'
