+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "postmodern",
-  "displayName": "postmodern"
+  "name": "Felfele",
+  "displayName": "Felfele"
 }
