+++ index.android.js
@@ -1,3 +1,3 @@
 import { AppRegistry } from 'react-native';
 import App from './src/App';
-AppRegistry.registerComponent('postmodern', () => App);
+AppRegistry.registerComponent('Felfele', () => App);
