+++ ios/Felfele/Images.xcassets/AppIcon.appiconset/ipad-40x40-1x.png
