+++ ios/Felfele/Images.xcassets/AppIcon.appiconset/ipad-40x40-2x.png
