+++ ios/Felfele/Images.xcassets/AppIcon.appiconset/ipad-50x50-1x.png
