+++ ios/Felfele/Images.xcassets/AppIcon.appiconset/ipad-50x50-2x.png
