+++ ios/Felfele/Images.xcassets/splash.imageset/icon-white-transparent-1x-1.png
