+++ ios/Felfele/Images.xcassets/splash.imageset/icon-white-transparent-2x-1.png
