+++ ios/Felfele/Images.xcassets/splash.imageset/icon-white-transparent-3x-1.png
