+++ ios/FelfeleTests/FelfeleTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface postmodernTests : XCTestCase
+@interface FelfeleTests : XCTestCase
 
 @end
 
-@implementation postmodernTests
+@implementation FelfeleTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
