+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "postmodern",
+  "name": "Felfele",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
